refresh

+++ aot/analysis/video_analysis/motion_energy/motion_energy.py
@@ -31,8 +31,22 @@ def get_motion_energy(video_file):
     return moten_features
 
 
+def check_result_existence(video):
+    if (result_path / f"{video}.pkl").exists():
+        print(f"Already processed: {video}")
+        return True
+    else:
+        # print(f"Processing: {video}")
+        return False
+
+
 all_videos = [video for video in os.listdir(stimuli_path) if video.endswith(".mp4")]
+
+
 for video in all_videos:
-    print(video)
-    moten_features = get_motion_energy(str(stimuli_path / video))
-    pickle.dump(moten_features, open(result_path / f"{video}.pkl", "wb"))
+    if check_result_existence(video):
+        continue
+    else:
+        print(video)
+        moten_features = get_motion_energy(str(stimuli_path / video))
+        pickle.dump(moten_features, open(result_path / f"{video}.pkl", "wb"))
